+++ shape/label.go
@@ -13,7 +13,7 @@ import (
 
 func NewLabel(text string) *Label {
 	return &Label{
-		Text:  template.HTMLEscapeString(text),
+		text:  template.HTMLEscapeString(text),
 		Font:  draw.DefaultFont,
 		Pad:   draw.DefaultPad,
 		class: "label",
@@ -24,7 +24,7 @@ type Label struct {
 	x int
 	y int
 
-	Text string
+	text string
 	href string
 
 	Font  draw.Font
@@ -32,8 +32,11 @@ type Label struct {
 	class string
 }
 
+func (l *Label) Text() string     { return l.text }
+func (l *Label) SetText(v string) { l.text = v }
+
 func (l *Label) String() string {
-	return fmt.Sprintf("label %s at %v", l.Text, &xy.Point{X: l.x, Y: l.y})
+	return fmt.Sprintf("label %s at %v", l.text, &xy.Point{X: l.x, Y: l.y})
 }
 
 func (l *Label) SetHref(v string) { l.href = v }
@@ -43,11 +46,11 @@ func (l *Label) SetX(x int)               { l.x = x }
 func (l *Label) SetY(y int)               { l.y = y }
 
 func (l *Label) Width() int {
-	return l.Font.TextWidth(longestLine(l.Text))
+	return l.Font.TextWidth(longestLine(l.text))
 }
 
 func (l *Label) Height() int {
-	return l.Font.LineHeight * (strings.Count(l.Text, "\n") + 1)
+	return l.Font.LineHeight * (strings.Count(l.text, "\n") + 1)
 }
 func (l *Label) Direction() Direction { return DirectionRight }
 func (l *Label) SetClass(c string)    { l.class = c }
@@ -61,7 +64,7 @@ func (l *Label) WriteSVG(out io.Writer) error {
 		w.Printf(`<a href="%s">`, l.href)
 	}
 	// support multilines
-	for i, line := range strings.Split(l.Text, "\n") {
+	for i, line := range strings.Split(l.text, "\n") {
 		if i > 0 { // write new line "after" each <text>, but not last
 			w.Print("\n")
 		}