+++ shape/state.go
@@ -52,7 +52,7 @@ func (r *State) title() *Label {
 		x:     r.x + r.Pad.Left,
 		y:     r.y + r.Pad.Top/2,
 		Font:  r.Font,
-		Text:  r.Title,
+		text:  r.Title,
 		class: "state-title",
 	}
 }