+++ shape/store.go
@@ -63,7 +63,7 @@ func (r *Store) title() *Label {
 		x:     r.x + r.Pad.Left,
 		y:     r.y + r.Pad.Top/2,
 		Font:  r.Font,
-		Text:  r.Title,
+		text:  r.Title,
 		class: r.class + "-title",
 	}
 }