+++ changelog.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.31.0] 2023-10-08
 
+- Hide field Label.Text
 - Remove aliased types and defaults in package shape
 - Hide field Diagram.Caption